tools/mass/openai-sdk-agents/requirements.txt
openai-agents>=0.1.0
pydantic>=2.0
